+++ src/main/java/com/imsweb/validation/ValidationEngine.java
@@ -94,7 +94,7 @@ public class ValidationEngine {
     /**
      * Engine version (used to check compatibility with the edits)
      */
-    private static final String _ENGINE_VERSION = "6.0";
+    private static final String _ENGINE_VERSION = "6.1";
 
     /**
      * The different context types supported by the engine