+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Version 2.1**
 
+- Increased the engine version to 6.1.
 - Changed implementation of MATCH and INLIST metafile context functions.
 - Fixed behavior of AT metafile context function.
 - Updated Groovy library from version 2.5.3 (indy) to version 2.5.4 (indy).